requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.2
numpy==2.2.3
pandas==2.2.3
pyyaml==6.0.2
scikit-learn==1.5.2
jinja2==3.1.4
